cleanup useTransactionFeeCalculation and useTransactionPriority
