+++ src/components/screens/send/form/formBtc.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import { MODULE_ASSETS } from '@constants';
 import { toRawLsk } from '@utils/lsk';
+import TransactionPriority, { useTransactionPriority, useTransactionFeeCalculation } from '@shared/transactionPriority';
 import FormBase from './formBase';
-import TransactionPriority from '../../../shared/transactionPriority';
 import useAmountField from './useAmountField';
-import useTransactionFeeCalculation from './useTransactionFeeCalculation';
-import useTransactionPriority from './useTransactionPriority';
 import useRecipientField from './useRecipientField';
 
 const txType = MODULE_ASSETS.transfer;
@@ -25,7 +23,7 @@ const FormBtc = (props) => {
     priorityOptions,
     token,
     account,
-    txData: {
+    transaction: {
       amount: toRawLsk(amount.value), txType, recipient: recipient.value,
     },
   });
