+++ src/components/screens/votingQueue/editor/editor.js
@@ -2,13 +2,12 @@ import React, { useMemo, useState } from 'react';
 
 import { tokenMap } from '@constants';
 import { toRawLsk } from '@utils/lsk';
+import TransactionPriority, { useTransactionFeeCalculation, useTransactionPriority } from '@shared/transactionPriority';
 import Box from '../../../toolbox/box';
 import BoxContent from '../../../toolbox/box/content';
 import BoxFooter from '../../../toolbox/box/footer';
 import { PrimaryButton } from '../../../toolbox/buttons';
-import TransactionPriority from '../../../shared/transactionPriority';
-import useTransactionFeeCalculation from '../../send/form/useTransactionFeeCalculation';
-import useTransactionPriority from '../../send/form/useTransactionPriority';
+
 import Table from '../../../toolbox/table';
 import ToggleIcon from '../toggleIcon';
 import VoteStats from '../voteStats';
@@ -118,7 +117,7 @@ const Editor = ({
     token,
     account,
     priorityOptions,
-    txData: {
+    transaction: {
       txType,
       nonce: account.nonce,
       senderPublicKey: account.publicKey,
