+++ src/components/shared/transactionPriority/transactionPriority.js
@@ -1,15 +1,17 @@
 import React, { useState, useMemo } from 'react';
 import PropTypes from 'prop-types';
-import { tokenMap, MODULE_ASSETS } from '@constants';
+
+import { tokenMap, MAX_ASSET_FEE } from '@constants';
 import {
   formatAmountBasedOnLocale,
 } from '@utils/formattedNumber';
 import { toRawLsk, fromRawLsk } from '@utils/lsk';
-import styles from './transactionPriority.css';
 import Input from '../../toolbox/inputs/input';
 import Icon from '../../toolbox/icon';
 import Tooltip from '../../toolbox/tooltip/tooltip';
 import Spinner from '../../toolbox/spinner';
+
+import styles from './transactionPriority.css';
 
 const CUSTOM_FEE_INDEX = 3;
 
@@ -46,14 +48,14 @@ const isCustomFeeValid = (value, hardCap, minFee) => {
 const TransactionPriority = ({
   t,
   token,
-  priorityOptions,
-  selectedPriority,
-  setSelectedPriority,
+  txType,
   fee,
   minFee,
   customFee,
   setCustomFee,
-  txType,
+  priorityOptions,
+  selectedPriority,
+  setSelectedPriority,
   className,
   loadError,
   isLoading,
@@ -61,9 +63,9 @@ const TransactionPriority = ({
   const [showEditIcon, setShowEditIcon] = useState(false);
   const [inputValue, setInputValue] = useState();
 
-  let hardCap = 0;
+  let maxFee = 0;
   if (token === tokenMap.LSK.key) {
-    hardCap = MODULE_ASSETS.getHardCap(txType);
+    maxFee = MAX_ASSET_FEE[txType];
   }
 
   const onClickPriority = (e) => {
@@ -84,7 +86,7 @@ const TransactionPriority = ({
     const newValue = e.target.value;
     if (token === tokenMap.LSK.key) {
       setInputValue(newValue);
-      if (isCustomFeeValid(newValue, hardCap, minFee)) {
+      if (isCustomFeeValid(newValue, maxFee, minFee)) {
         setCustomFee({ value: newValue, feedback: '', error: false });
       } else {
         setCustomFee({ value: undefined, feedback: 'invalid custom fee', error: true });
@@ -134,10 +136,13 @@ const TransactionPriority = ({
             if (index === 0) {
               priority.title = priority.value === 0 ? 'Normal' : 'Low';
             } else if (index === 3) {
-              disabled = priority.value === 0 && !loadError; // Custom fee option
+              // Custom fee option
+              disabled = priority.value === 0 && !loadError;
             } else {
-              disabled = priority.value === 0 || loadError; // Medium and high fee option
+              // Medium and high fee option
+              disabled = priority.value === 0 || loadError;
             }
+
             return (
               <button
                 key={`fee-priority-${index}`}
@@ -187,8 +192,8 @@ const TransactionPriority = ({
                 onChange={onInputChange}
                 onBlur={onInputBlur}
                 onFocus={onInputFocus}
-                status={!isCustomFeeValid(inputValue, hardCap, minFee) ? 'error' : 'ok'}
-                feedback={`fee must be between ${minFee} and ${fromRawLsk(hardCap)}`}
+                status={!isCustomFeeValid(inputValue, maxFee, minFee) ? 'error' : 'ok'}
+                feedback={`fee must be between ${minFee} and ${fromRawLsk(maxFee)}`}
               />
             ) : (
               <span className={`${styles.feeValue} fee-value`} onClick={onClickCustomEdit}>
