+++ src/components/shared/transactionPriority/useTransactionFeeCalculation.js
@@ -1,67 +1,9 @@
 import { useSelector } from 'react-redux';
 import { useEffect, useReducer } from 'react';
-import { tokenMap, minAccountBalance } from '@constants';
 import {
   getTransactionFee,
 } from '@utils/api/transaction';
-import { toRawLsk } from '@utils/lsk';
-
-const calculateAvailableBalance = (balance, token) => {
-  if (token !== tokenMap.LSK.key) {
-    return balance;
-  }
-  if (balance <= minAccountBalance) {
-    return balance;
-  }
-  return balance - minAccountBalance;
-};
-
-const initialFee = {
-  value: 0,
-  error: false,
-  feedback: '',
-};
-
-const getInitialState = account => ({
-  fee: initialFee,
-  minFee: initialFee,
-  maxAmount: {
-    value: account.balance,
-    error: false,
-    feedback: '',
-  },
-});
-
-const actionTypes = {
-  setFee: 'SET_FEE',
-  setMinFee: 'SET_MIN_FEE',
-  setMaxAmount: 'SET_MAX_AMOUNT',
-};
-
-const reducer = (state, action) => {
-  switch (action.type) {
-    case actionTypes.setFee:
-      return { ...state, fee: action.payload.response };
-
-    case actionTypes.setMinFee:
-      return { ...state, minFee: action.payload.response };
-
-    case actionTypes.setMaxAmount: {
-      const balance = action.payload.account.balance;
-      const token = action.payload.token;
-      const availableBalance = calculateAvailableBalance(balance, token);
-      const result = {
-        ...action.response,
-        value: availableBalance - toRawLsk(action.params.value),
-      };
-
-      return { ...state, ...result };
-    }
-
-    default:
-      throw Error(`reducer not implemented for ${action}`);
-  }
-};
+import { actionTypes, reducer, getInitialState } from './reducer';
 
 const useTransactionFeeCalculation = ({
   selectedPriority, transaction, token, account, priorityOptions,
